Introduce --stdin parameter
Depending on the particular use-case it might be more suitable to
retrieve the list of files to check from the program arguments or from
stdin.
This change introduces the --stdin parameter that causes the script to
read the files to work on from stdin as opposed to from the argument
vector supplied to it.