Add filter for Rust files
Until now we were only able to filter Python files. This change adds
support for filtering out Rust files as well. In order to differentiate
between the two cases two newly introduced parameters, --python and
--rust, can be used.
Detection of Rust files is slightly easier than detection of Python
files as they are strictly compiled. For that reason we do not need to
look inside a file but use the extension as the only guide for the
classification.