Introduce -v/--verbose parameter
The file-filter script may be used in arbitrarily complex architectures
where direct investigation of a problem may not be possible simply
because hooking into a specific part of it is not easily feasible. This
change introduces a new parameter, -v/--verbose, that may or may not
help in such situations. When set, this paramter causes the script to
print the resulting set of files to stderr.