Ignore directories and symbolic links
This change causes the file filter script to ignore directories and
symbolic links when filtering files. Directories are to be ignored for
the obvious reason of not being a file in the strict sense. Symbolic
links are generally not of interest because it is only their target that
is of interest. This target is either contained in the list itself, in
which case it will be handled correctly, or missing, in which case we
argue that it is of no interest.